homedeyattayo

+++ app/Http/Controllers/TasksController.php
@@ -17,11 +17,26 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        // $tasks = Task::all();
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        // return view('tasks.index', [
+        //     'tasks' => $tasks,
+        // ]);
+        
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            // $data += $this->counts($user);
+            return view('tasks.index', $data);
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -51,12 +66,17 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status; 
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,  
+        ]);
+        
+        // $task = new Task;
+        // $task->status = $request->status; 
+        // $task->content = $request->content;
+        // $task->save();
 
-        return redirect('/');
+        return redirect('/tasks');
     }
 
     /**
@@ -68,11 +88,20 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        
+        if (\Auth::user()->id === $task->user_id){
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+            
+           return redirect('/');
+        }
     }
+
+    //     return view('tasks.show', [
+    //         'task' => $task,
+    //     ]);
+    // }
 
     /**
      * Show the form for editing the specified resource.
@@ -83,10 +112,18 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
+        
+        if (\Auth::user()->id === $task->user_id){
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+            
+           return redirect('/');
+        }
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        // return view('tasks.edit', [
+        //     'task' => $task,
+        // ]);
     }
 
     /**
@@ -120,8 +157,16 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        
+        if (\Auth::user()->id === $task->user_id){
+            $task->delete();
+            return redirect('/');
+        } else {
+            return view('welcome');
+        }
+        
+        // $task->delete();
 
-        return redirect('/');
+        // return redirect('/');
     }
 }
